myapi/requirements/prod.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-http-client==3.3.7
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3

# Dev-only dependencies
python-dotenv==1.1.1
django-debug-toolbar==4.3.0
ipython
coverage
pytest
pytest-django
django-extensions
